backend/requirements.txt
# PostgreSQL-only requirements.txt - No C++ build tools needed

# Flask and core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# PostgreSQL Database - Using pg8000 instead of psycopg2 (pure Python)
pg8000==1.30.2

# Password hashing - Using passlib instead of bcrypt (pure Python)
passlib==1.7.4

# Environment and utilities
python-dotenv==1.0.0
requests==2.31.0

# AI/ML APIs
openai==1.3.0

# Image processing
Pillow==10.0.1

# JSON Web Tokens
PyJWT==2.8.0